requirements.txt
image
opencv-python
matplotlib
numpy
Pillow
pycocotools
scipy
six
tensorboardX
torch
torchvision
tqdm
torchfcn
imgviz
labelme